Update OrderRequestScreen to Integrate Cart and Order Request Logic
- Integrated CartCubit and OrderRequestCubit into the OrderRequestScreen to dynamically fetch and display cart items and their total prices.
- Modified the FatoraDetailsRow widget to accept double values for fatoraCount to support displaying total prices.
- Updated the OrderRequestItem widget to display cart item details, including product images, names, and prices, fetched from the CartCubit.
- Enhanced the PayRow widget to dynamically calculate and display the total price including shipping costs, based on the current cart state.
- Removed the const constructor from OrderRequestScreen to allow dynamic initialization of cart items.

+++ lib/feature/order_request/presentation/view/widgets/fatora_details_row.dart
@@ -2,8 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class FatoraDetailsRow extends StatelessWidget {
-  const FatoraDetailsRow({super.key, this.fatoraCount, this.fatoraTitle});
-  final int? fatoraCount;
+  const FatoraDetailsRow({
+    super.key,
+    this.fatoraCount,
+    this.fatoraTitle,
+  });
+  final double? fatoraCount;
   final String? fatoraTitle;
 
   @override
